:sparkles: Add interface to load commands from FS
This allows us to implementing loading from
not just single YAML files, but from directories and skip yaml template files and all that jazz.

Useful for both geppetto and escuse-me who don't play too well with YAML in YAML.

In general, this is probably the right thing for sqleton too, because we would probably get syntax highlighting
from our editors.